WIP aspnet webhook

+++ src/Yandex.Messenger.Bot.Sdk/Models/Requests/GetUpdateRequest.cs
@@ -4,5 +4,5 @@ public record GetUpdateRequest
 {
     public int Limit { get; init; } = 100;
 
-    public int Offset { get; init; } = 0;
+    public long Offset { get; init; } = 0;
 }
